Priorioties component  performance enhacement
